src/third_party/blink/web_tests/clipboard/async-clipboard/permissions/writeText-denied.https.html

<!doctype html>
<meta charset="utf-8">
<title>navigator.clipboard.writeText() fails when permission denied</title>
<link rel="help" href="https://w3c.github.io/clipboard-apis/#async-clipboard-api">
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../http/tests/resources/permissions-helper.js"></script>
<script>
promise_test(async t => {
  await PermissionsHelper.setPermission('clipboard-sanitized-write', 'denied');
  await promise_rejects_dom(t, 'NotAllowedError', navigator.clipboard.writeText('xyz'));
}, 'navigator.clipboard.writeText() fails when permission denied');
</script>
